:lipstick: design : detail 페이지까지의 설정 완료
