+++ src/app/quiz/[category]/[number]/page.tsx
@@ -1,5 +1,5 @@
 const quizDetailPage = () => {
-  return <div>category</div>
+  return <div>detail 문제 페이지</div>
 }
 
 export default quizDetailPage
