+++ src/components/question/question.tsx
@@ -12,10 +12,10 @@ export interface QuestionProps {
 }
 
 export const Question: React.FC<QuestionProps> = ({ question }) => {
-  const { category_id } = question
+  const { category_id, id } = question
 
   return (
-    <Link href={`/quiz/${category_id}`}>
+    <Link href={`/quiz/${category_id}/${id}`}>
       <div className="border p-4 my-4">
         <div className="text-lg text-primaryColor font-semibold">{question.category_id}</div>
         <div className="text-lg font-semibold">{question.question}</div>
